requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
redis==5.0.1
bcrypt==4.1.2
cryptography==41.0.7
requests==2.31.0
psycopg2-binary==2.9.9
email-validator==2.1.0
bleach==6.1.0
clerk-sdk==0.1.0  # Add Clerk SDK for authentication
